Jia Hong part resolved
